fix(sync): move decryptTokenEdge to separate file to resolve Edge webpack build error
